backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20

pandas==2.3.1
pyarrow==21.0.0
orjson==3.11.1
networkx==3.4.2
numpy==2.2.6

langchain==0.3.27
langchain-openai==0.3.28
openai==1.99.9
tiktoken==0.10.0

httpx==0.28.1
tenacity==9.1.2
python-dotenv==1.1.1
PyYAML==6.0.2

pytest==8.4.1
pytest-asyncio==1.1.0
ruff==0.12.5
google-genai==1.31.0
